fix: allow non API in args
Running bashdoc without -A flag return no content. This
patch solve that issue. If not API is provided all documentation
blocks as considered part of the same default API.

Also add some improvements in readability in the provided CSS,
and fix test.

Thanks to Alan Transon for point me the missing API bug!

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>